further adjustments to game page component allocation
