+++ src/Game/GamePage.js
@@ -1,13 +1,29 @@
-import { useMemo, useState } from "react";
+import { useMemo, useState, useEffect } from "react";
 import "./Game.css";
 import AttackButton from "./AttackButton";
+import { supabase } from "../supabase_client";
 import AttackTargetSelect from "./AttackTargetSelect";
 
 const GamePage = (props) => {
   const { roomCode, playerName } = props;
   const [totalPlayerData, setTotalPlayerData] = useState([]);
   const [attackTarget, setAttackTarget] = useState("Select a target");
+  const [currentTurnPlayerIndex, setCurrentTurnPlayerIndex] = useState(0);
 
+  useEffect(() => {
+    const fetchHitpoints = async () => {
+      const { data } = await supabase
+        .from("Players")
+        .select("name, Hitpoints")
+        .order("id", { ascending: true })
+        .eq("roomID", roomCode);
+      const playerHp = data.map((playerData) => {
+        return { name: playerData.name, Hitpoints: playerData.Hitpoints };
+      });
+      setTotalPlayerData(playerHp);
+    };
+    fetchHitpoints();
+  }, []);
   const myPlayerData = useMemo(() => {
     return totalPlayerData.find((playerData) => playerData.name === playerName);
   }, [totalPlayerData]);
@@ -26,6 +42,46 @@ const GamePage = (props) => {
   };
 
   const hasWon = useMemo(checkWinCondition, [totalPlayerData]);
+
+  const updateHitpointValues = (currentState, payload) => {
+    const updateHitpoints = currentState.map((playerData) => {
+      if (payload.new.name === playerData.name) {
+        return {
+          name: payload.new.name,
+          Hitpoints: payload.new.Hitpoints > 0 ? payload.new.Hitpoints : 0,
+        };
+      } else if (payload.table === "Players") {
+        return playerData;
+      }
+    });
+    return updateHitpoints;
+  };
+
+  useEffect(() => {
+    let mySubscription = supabase
+      .from("Players")
+      .on("UPDATE", (payload) => {
+        setTotalPlayerData((currentState) =>
+          updateHitpointValues(currentState, payload)
+        );
+      })
+      .subscribe();
+    return () => {
+      supabase.removeSubscription(mySubscription);
+    };
+  }, []);
+
+  useEffect(() => {
+    let turnSubscription = supabase
+      .from("Rooms")
+      .on("UPDATE", (payload) => {
+        setCurrentTurnPlayerIndex(payload.new.turnNumber);
+      })
+      .subscribe();
+    return () => {
+      supabase.removeSubscription(turnSubscription);
+    };
+  }, []);
 
   return (
     <div>
@@ -46,6 +102,7 @@ const GamePage = (props) => {
             playerName={playerName}
             isPlayerEliminated={isPlayerEliminated}
             setTotalPlayerData={setTotalPlayerData}
+            currentTurnPlayerIndex={currentTurnPlayerIndex}
           />
           {isPlayerEliminated === true && <h3>You are dead.</h3>}
           {hasWon === true && <h3>You have won!</h3>}
